Made initialization from file/input. Calc_isomif not working??  (buffer overflow detected)

+++ IsoMIF.py
@@ -1,3 +1,5 @@
+# %load IsoMIF.py
+
 """Using 2 pdb ids and numbers of two clefts from the biggest gets tanimoto coef between them"""
 
 import os
@@ -6,15 +8,14 @@ import requests
 from pathlib import Path  # Library for easier paths processing
 import mmap
 
-
-global ROOT_PATH, GSL_PATH, GET_CLEFT_PATH, ISOMIF_PATH, REDUCE_PATH, SYSTEM_NAME
+global GSL_PATH, GET_CLEFT_PATH, ISOMIF_PATH, REDUCE_PATH, SYSTEM_NAME
 global HIVE_PATH, PDB_PATH, MIF_NAME, ISOMIF_NAME, N_cavities
 
 PDB_PATH = None
 N_cavities = 1
 
 
-def make_dir(dirList):
+def make_dir_from_list(dirList):
     """Make directories with paths from dirList"""
     for dirName in dirList:
         if not os.path.exists(dirName):
@@ -25,14 +26,15 @@ def make_dir(dirList):
             #print("Directory " , dirName ,  " already exists")
 
 
-def make_hive(root_path):
+def make_hive(root_isomif):
     """Create all needed for IsoMIF data storage directories"""
     List = []
-    List.append(str(Path(root_path) / 'hive'))
+    List.append(str(Path(root_isomif)))
+    List.append(str(Path(root_isomif) / 'hive'))
     for last_name in ('clefts', 'match', 'matchView',
                       'mifs', 'mifView', 'pdb'):
-        List.append(str(Path(root_path) / 'hive' / last_name))
-    make_dir(List)
+        List.append(str(Path(root_isomif) / 'hive' / last_name))
+    make_dir_from_list(List)
 
 
 def download_url(url, path=None, name=None):
@@ -41,7 +43,7 @@ def download_url(url, path=None, name=None):
     if path:
         paths = []
         paths.append(path)
-        make_dir(paths)
+        make_dir_from_list(paths)
         open(os.path.join(paths[0], name), 'wb').write(r.content)
     return r.content.decode('utf-8')
 
@@ -62,23 +64,50 @@ def download_pdb(pdb, pdb_path=None):
         download_url(url, str(path_uniprot), name)
 
 
-def isomif_init(root_path, gsl_path, get_cleft_path,
-                isomif_path, reduce_path, system_name, n_cavities):
-    """Prepare everything  for the IsoMIF work
-    Input -- strings
+def isomif_init(root, gsl_path=None, reduce_path=None,
+                system_name='linux_x86_64', n_cavities=1, 
+                overwrite_init=False):
+    """ Prepare everything (paths, default settings)  for the IsoMIF work. 
+    If only root is in input, then try to find root/init.txt and get settings from it
+    
+    Input -- strings:
+            root - Root for making storage of all needed for Isomif data
+            gsl_path - Path to GSL library with 'lib' and 'include' subdirectories
+            reduce_path - Path to reduce program
+            system_name - string of system name, could be 'linux_x86_64' (default) or 'mac_x86_64'
+            
+        n_cavities - Default number of cavities to find
+        boolean overwrite_init -- True if needed to overwrite old init.txt
     """
-    global ROOT_PATH, GSL_PATH, GET_CLEFT_PATH, ISOMIF_PATH, REDUCE_PATH, SYSTEM_NAME
+    global GSL_PATH, GET_CLEFT_PATH, ISOMIF_PATH, REDUCE_PATH, SYSTEM_NAME
     global HIVE_PATH, PDB_PATH, MIF_NAME, ISOMIF_NAME, N_cavities
-    # Further paths needed to define by user
+    
+    path_init = str(Path(root)  / 'Isomif' / 'init.txt')
+    # If need to initialize from input and create/overwrite init file 
+    if not Path(path_init).is_file() or overwrite_init:
+        with open(str(Path(root) / 'Isomif' / 'init.txt'), 'w+') as init_file:
+            init_file.write('\n'.join([root, gsl_path, reduce_path, 
+                                  system_name, str(n_cavities)]))
+    # If init.txt exists and we don't need to overwrite it
+    else:
+        with open(Path(root) / 'Isomif' / 'init.txt', 'r') as init_file:
+            lines = init_file.readlines()
+            root = lines[0].rstrip()
+            gsl_path = lines[1].rstrip()
+            reduce_path = lines[2].rstrip()
+            system_name = lines[3].rstrip()
+            n_cavities = int(lines[4])
+
+    # Further global paths needed to be defined
     #
-    # Root for making storage of produced data    
-    ROOT_PATH = root_path
+    # Root for making storage of all needed for Isomif data    
+    root_isomif = str(Path(root) / 'Isomif')
     # Path to GSL library with 'lib' and 'include' subdirectories
     GSL_PATH = gsl_path
-    # Path to Get_Cleft directory
-    GET_CLEFT_PATH = get_cleft_path
-    # Path to Isomif directory
-    ISOMIF_PATH = isomif_path
+    # Path to Get_Cleft-master directory
+    GET_CLEFT_PATH = str(Path(root_isomif) / 'Get_Cleft-master')
+    # Path to Isomif-master directory
+    ISOMIF_PATH = str(Path(root_isomif) / 'IsoMif-master')
     # Path to reduce program
     REDUCE_PATH = reduce_path
     # System name, could be 'linux_x86_64' or 'mac_x86_64'
@@ -87,32 +116,35 @@ def isomif_init(root_path, gsl_path, get_cleft_path,
     N_cavities = n_cavities
 
     # Where all data stored
-    HIVE_PATH = str(Path(ROOT_PATH) / 'hive')
-    PDB_PATH = str(Path(ROOT_PATH) / 'hive' / 'pdb')
+    HIVE_PATH = str(Path(root_isomif) / 'hive')
+    PDB_PATH = str(Path(root_isomif) / 'hive' / 'pdb')
 
     MIF_NAME = 'mif_' + SYSTEM_NAME + '_compiled'
     ISOMIF_NAME = 'isomif_' + SYSTEM_NAME + '_compiled'
 
     # Creating needed for IsoMIF directories inside ROOT
-    make_hive(ROOT_PATH)
-
+    make_hive(root_isomif)
+        
 
 def isomif_compile():
-    """Compile Get_Cleft for this OS. Initialisation needed before this"""
+    """Compile Get_Cleft for this OS. Initialisation by isomif_init function is needed before this"""
     subprocess.check_output(['gcc', str(Path(GET_CLEFT_PATH) / 'Get_Cleft.c'),
-                             '-o', 'Get_Cleft',
+                             '-o', str(Path(GET_CLEFT_PATH) / 'Get_Cleft'),
                              '-O3', '-lm',
                              ])
+    print('Get_Cleft compiled')
     # Compile MIF and IsoMIF for linux_x86_64, needs path to gsl
     subprocess.check_output(['g++', str(Path(ISOMIF_PATH) / 'mif.cpp'),
                             '-o', str(Path(ISOMIF_PATH) / MIF_NAME),
                             '-O3', '-lm',
                             ])
+    print('MIF compiled')
     subprocess.check_output(['g++', str(Path(ISOMIF_PATH) / 'isomif.cpp'),
                             '-o', str(Path(ISOMIF_PATH) / ISOMIF_NAME),
                             '-O3', '-lm', '-lgsl', '-lgslcblas', '-L', str(Path(GSL_PATH) / 'lib'),
                             '-I', str(Path(GSL_PATH) / 'include'),
                             ])
+    print('IsoMIF compiled')
 
     
 def get_cavities(pdb, n_cavities=N_cavities, pdb_path=None):
@@ -130,36 +162,53 @@ def add_hydrogens_by_reduce(pdb, pdb_path=None):
     """Add hydrogens to .pdb using reduce, save in the same directory adding 'h' to pdb name"""
     if pdb_path is None:
         pdb_path = PDB_PATH
-    subprocess.check_output(REDUCE_PATH + \
-                            ' -p ' + str(Path(pdb_path) / (pdb + '.pdb')) + \
-                            '  >' + str(Path(pdb_path) / (pdb + 'h.pdb')), shell=True
-                            )
+    # Check if hydrogens were added before
+    if not Path(str(Path(pdb_path) / (pdb + 'h.pdb'))).is_file():
+        print(str(Path(pdb_path) / (pdb + '.pdb')))
+        subprocess.check_output(REDUCE_PATH + \
+                                ' -p ' + str(Path(pdb_path) / (pdb + '.pdb')) + \
+                                ' > ' + str(Path(pdb_path) / (pdb + 'h.pdb')), shell=True
+                                )  # Somehow with list of parameters does't work properly
+        print(f'Added hydrogens to {pdb}.pdb')
+    else:
+        print(f'Were before added hydrogens to {pdb}.pdb')
 
 
-def calc_mif(pdb,  n_cavities=N_cavities, pdb_path=None, res_mif=1, vis=False):
+def calc_mif(pdb,  n_cavities=N_cavities, pdb_path=None, res_mif=1, vis=True):
     """Get MIFs for n_cavities of pdb placed in pdb_path. Vis=True to make visualisation
     res -- Resolution
     0 => 2 Ang, 1 => 1.5 Ang, 2 => 1.0 Ang, 3 => 0.5 Ang"""
     if pdb_path is None:
         pdb_path = PDB_PATH
+    f_calculated_something = False
     for i in range(1, n_cavities + 1):
-        subprocess.check_output([str(Path(ISOMIF_PATH) / MIF_NAME),
-                                '-p', str(Path(PDB_PATH) / (pdb + 'h.pdb')),
-                                '-g', str(Path(HIVE_PATH) / 'clefts' / (pdb + '_sph_' + str(i) + '.pdb')),
-                                '-o', str(Path(HIVE_PATH) / 'mifs'),
-                                'z', str(res_mif),
-                                ])
-        # Rename files, as by default makes just smth like 1E8Xh.mif and 1E8Xh_cpy.pdb despite number of cleft
-        os.rename(str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h.mif')), 
-                  str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '.mif')))
-        os.rename(str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_cpy.pdb')), 
-                  str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '_cpy.pdb')))
-        # Produce file for visualisation
-        if vis:
-            subprocess.check_output(['perl', str(Path(ISOMIF_PATH) / 'mifView.pl'),
-                                     ' -m ', str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h.mif')),
-                                     ' -o ', str(Path(HIVE_PATH) / 'mifView'), #/ (pdb + str(i))
+        # Check if MIF is already calculated
+        if not Path(str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '.mif'))).is_file():
+            f_calculated_something = True
+            subprocess.check_output([str(Path(ISOMIF_PATH) / MIF_NAME),
+                                    '-p', str(Path(PDB_PATH) / (pdb + 'h.pdb')),
+                                    '-g', str(Path(HIVE_PATH) / 'clefts' / (pdb + '_sph_' + str(i) + '.pdb')),
+                                    '-o', str(Path(HIVE_PATH) / 'mifs'),
+                                    'z', str(res_mif),
                                     ])
+            # Rename files, as by default makes just smth like 1E8Xh.mif and 1E8Xh_cpy.pdb despite number of cleft
+            os.rename(str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h.mif')), 
+                      str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '.mif')))
+            os.rename(str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_cpy.pdb')), 
+                      str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '_cpy.pdb')))
+            # Produce file for visualisation
+            if vis:
+                subprocess.check_output(['perl', str(Path(ISOMIF_PATH) / 'mifView.pl'),
+                                         ' -m ', str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '.mif')),
+                                         ' -o ', str(Path(HIVE_PATH) / 'mifView'), #/ (pdb + str(i))
+                                    ])
+                print('perl' + str(Path(ISOMIF_PATH) / 'mifView.pl')+ \
+                                         ' -m ' + str(Path(HIVE_PATH) / 'mifs' / (pdb + 'h_' + str(i) + '.mif')) + \
+                                         ' -o ' + str(Path(HIVE_PATH) / 'mifView'))
+        else:
+            print(f'Was before calculated MIF for {i}th(st, nd)')
+    if f_calculated_something:
+        print(f'Calculated MIFs of {n_cavities} the biggest cavities in {pdb}h.pdb')
             
 
 def calc_isomif(pdb1, pdb2, i1=1, i2=1, res_isomif=1, node_variab=2.0, vis=False, res_nodes=1):
@@ -167,18 +216,34 @@ def calc_isomif(pdb1, pdb2, i1=1, i2=1, res_isomif=1, node_variab=2.0, vis=False
     in HIVE_PATH / 'mifs'
     """
     # SOMETIMES COLLAPSES WHEN NOT ENOUGH MEMORY
-    subprocess.check_output([str(Path(ISOMIF_PATH) / ISOMIF_NAME),
-                            '-p1', str(Path(HIVE_PATH) / 'mifs' / (pdb1 + 'h_' + str(i1) + '.mif')),
-                            '-p2', str(Path(HIVE_PATH) / 'mifs' / (pdb2 + 'h_' + str(i2) + '.mif')),
-                            '-o', os.path.join(str(Path(HIVE_PATH) / 'match'), ''),
-                            '-c', str(res_isomif), '-d', str(node_variab),
-                            ])
-    if vis:
-        subprocess.check_output(['perl', str(Path(ISOMIF_PATH) / 'isoMifView.pl'),
-                                 '-m', str(Path(HIVE_PATH) / 'match' / (pdbh1 + '_match_' + pdbh2 + '.isomif')),
-                                 '-o', os.path.join(str(Path(HIVE_PATH) / 'matchView'), ''),
-                                 '-g', str(res_nodes)
+    pdbh1 = pdb1 + 'h'
+    pdbh2 = pdb2 + 'h'
+    # Check if IsoMIF isn't calculated already
+    if not Path(str(Path(HIVE_PATH) / 'match' / (pdbh1 + '_' + str(i1) + '_match_' + \
+                                                 pdbh2 + '_' + str(i2) + '.isomif'))).is_file():
+        print(' '.join([str(Path(ISOMIF_PATH) / ISOMIF_NAME),
+                                '-p1', str(Path(HIVE_PATH) / 'mifs' / (pdbh1 + '_' + str(i1) + '.mif')),
+                                '-p2', str(Path(HIVE_PATH) / 'mifs' / (pdbh2 + '_' + str(i2) + '.mif')),
+                                '-o', os.path.join(str(Path(HIVE_PATH) / 'match'), ''),
+                                '-c', str(res_isomif), '-d', str(node_variab),
+                                ]))
+        subprocess.check_output([str(Path(ISOMIF_PATH) / ISOMIF_NAME),
+                                '-p1', str(Path(HIVE_PATH) / 'mifs' / (pdbh1 + '_' + str(i1) + '.mif')),
+                                '-p2', str(Path(HIVE_PATH) / 'mifs' / (pdbh2 + '_' + str(i2) + '.mif')),
+                                '-o', os.path.join(str(Path(HIVE_PATH) / 'match'), ''),
+                                '-c', str(res_isomif), '-d', str(node_variab),
                                 ])
+        
+        if vis:
+            subprocess.check_output(['perl', str(Path(ISOMIF_PATH) / 'isoMifView.pl'),
+                                     '-m', str(Path(HIVE_PATH) / 'match' / (pdbh1 + '_' + str(i1) + \
+                                                                '_match_' + pdbh2 + '_' + str(i2) + '.isomif')),
+                                     '-o', os.path.join(str(Path(HIVE_PATH) / 'matchView'), ''),
+                                     '-g', str(res_nodes)
+                                    ])
+        print(f'Calculated IsoMIF for {i1}th(st, nd) and {i2}th(st, nd) biggest cavities in {pdb1}h.pdb and {pdb2}h.pdb')
+    else:
+        print(f'Was before calculated IsoMIF for {i1}th(st, nd) and {i2}th(st, nd) biggest cavities in {pdb1}h.pdb and {pdb2}h.pdb') 
         
         
 def get_tanimoto_from_isomif_file(pdb1, pdb2, i1, i2):
@@ -199,22 +264,23 @@ def mif_from_pdb_simple(pdb, n_cavities=N_cavities):
     download_pdb(pdb, PDB_PATH)
     add_hydrogens_by_reduce(pdb)
     get_cavities(pdb, n_cavities)
-    calc_mif(pdb, n_cavities)
+    calc_mif(pdb, n_cavities, vis=True)
 
 
+def calc_mifs_of_whole_drugbank():
+    """"""
 # Initialisation of paths and default parameters
 # (maybe, it's worth to put initialisation parameters into .txt?)
+root = '/home/anton_maximov/BACHELOR'
+root_isomif = str(Path(root) / 'Isomif')
+#root = '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR/Isomif'
 isomif_init(
     # Root directory
-    '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR',
+    root,
     # GSL directory
-    '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR/gsl',
-    # Get_Cleft directory
-    '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR/Get_Cleft-master',
-    # IsoMif directory
-    '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR/IsoMif-master',
-    # Reduce executive,
-    '/media/anton/b8150e49-6ff0-467b-ad66-40347e8bb188/anton/BACHELOR/IsoMif-master/reduce.3.23.130521',
+    '/home/anton_maximov/gsl',
+    # Reduce executive
+    os.path.join(root_isomif, 'reduce.3.23.130521'),
     # System name, could be 'linux_x86_64' or 'mac_x86_64'
     'linux_x86_64', 
     # Default number of cavities to find
@@ -224,18 +290,12 @@ isomif_init(
 # Compilation if hasn't been done before
 #isomif_compile()
 
-# Example of getting pdbs from uniprot:
+# Example of getting pdbs from uniprot: 
 #uniprot1 = 'P00734'  # Prothrombin
 #uniprot2 = 'P00736'  # Cetuximab
 #pdb1 = get_pdbs_from_uniprot(uniprot1, PDB_PATH)[1]
 #pdb2 = get_pdbs_from_uniprot(uniprot2, PDB_PATH)[2]
-
-pdb1 = input()
-pdb2 = input()
-i1 = input()
-i2 = input()
-
-mif_from_pdb_simple(pdb1, i1)
-mif_from_pdb_simple(pdb2, i2)
-calc_isomif(pdb1, pdb2, i1, i2)
-return get_tanimoto_from_isomif_file(pdb1, pdb2, i1, i2)
+#mif_from_pdb_simple(pdb1, i1)
+#mif_from_pdb_simple(pdb2, i2)
+#calc_isomif(pdb1, pdb2, i1, i2) DOESN'T WORK
+#return get_tanimoto_from_isomif_file(pdb1, pdb2, i1, i2)
